% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfuncts.R
\name{anualidad_n}
\alias{anualidad_n}
\title{Capitalización anual}
\usage{
anualidad_n(deposito, interes, periodicidad, tiempo_anual)
}
\arguments{
\item{deposito}{Valor de los depósitos constantes que se deben hacer en el periodo dado}

\item{interes}{Valor de la tasa de interés a lo largo del periodo}

\item{periodicidad}{Unidades de tiempo (días, meses, años) en las que se dejará la inversión actual}

\item{tiempo_anual}{Número de periodos distribuidos con amplitud anual}
}
\value{

}
\description{
Capitalización anual
}
\details{
La capitalización consiste en depositar de manera constante un capital, el cuál irá aumentando y generando un monto final con una ganancia por una tasa de interés dada. La función regresa únicamente el valor numérico. Útil cuando se quieren hacer cálculos con ese valor.
}
\examples{
anualidad_n(20000,0.12,2,4.5)
}
