% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfuncts.R
\name{capital_compuesto}
\alias{capital_compuesto}
\title{Cálculo del capital}
\usage{
capital_compuesto(tiempo, valor_final, periodicidad, interes)
}
\arguments{
\item{tiempo}{Unidades de tiempo (días, meses, años) en las que se dejará la inversión actual}

\item{periodicidad}{Amplitud de las unidades de tiempo. Ej: 1 equivale a un año, 2 a un semestre, 12 a un mes, 365 a un día.}

\item{interes}{Tasa de interés compuesta a la que se dejará la inversión inicial}

\item{monto}{Monto esperado a recibir al final del plazo}
}
\value{

}
\description{
Cálculo del capital
}
\details{
Función para calcular el capital necesario para obtener un monto dado a un cierto tiempo, con una tasa de interés compuesta y capitalización dadas. Regresa una cadena de texto explicando el valor necesario.
}
\examples{
capital_compuesto(12,150,12,0.12)
}
