% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfuncts.R
\name{interes_compuesto}
\alias{interes_compuesto}
\title{Cálculo de la tasa de interés compuesta.}
\usage{
interes_compuesto(capital, tiempo, monto, periodicidad)
}
\arguments{
\item{capital}{Valor de la inversión inicial}

\item{tiempo}{Unidades de tiempo (días, meses, años) en las que se dejará la inversión actual}

\item{monto}{Monto esperado a recibir al final del plazo}

\item{periodicidad}{Amplitud de las unidades de tiempo. Ej: 1 equivale a un año, 2 a un semestre, 12 a un mes, 365 a un día.}
}
\value{

}
\description{
Cálculo de la tasa de interés compuesta.
}
\details{
Se calcula la tasa de interés compuesta necesaria para obtener un monto a un cierto periodo, con una inversión inicial dada. Devuelve una cadena de texto con la tasa de interés necesaria.
}
\examples{
interes_compuesto(100,12,50,12)
}
