% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfuncts.R
\name{interes_simple_n}
\alias{interes_simple_n}
\title{Cálculo de la tasa de interés}
\usage{
interes_simple_n(capital, tiempo, monto, periodicidad)
}
\arguments{
\item{capital}{Valor de la inversión inicial}

\item{tiempo}{Unidades de tiempo (días, meses, años) en las que se dejará la inversión actual}

\item{monto}{Monto esperado a recibir al final del plazo}

\item{periodicidad}{Amplitud de las unidades de tiempo. Ej: 1 equivale a un año, 2 a un semestre, 12 a un mes, 365 a un día.}
}
\value{

}
\description{
Cálculo de la tasa de interés
}
\details{
Se calcula la tasa de interés necesaria para obtener un monto a un cierto periodo, con una inversión inicial dada. Unicamente devuelve el valor numérico. Útil cuando se requieren hacer cálculos adicionales con ese valor.
}
\examples{
interes_simple_n(100,12,50,12)
}
