% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfuncts.R
\name{monto_compuesto_n}
\alias{monto_compuesto_n}
\title{Cálculo del monto generado a una tasa compuesta.}
\usage{
monto_compuesto_n(capital, tiempo, interes, periodicidad)
}
\arguments{
\item{capital}{Valor de la inversión inicial}

\item{tiempo}{Unidades de tiempo (días, meses, años) en las que se dejará la inversión actual}

\item{interes}{Tasa de interés compuesta a la que se dejará la inversión inicial}

\item{periodicidad}{Amplitud de las unidades de tiempo. Ej: 1 equivale a un año, 2 a un semestre, 12 a un mes, 365 a un día.}
}
\value{

}
\description{
Cálculo del monto generado a una tasa compuesta.
}
\details{
La función calcula el monto que va a generar nuestra inversión a una tasa de interés compuesta dada, con un periodo establecido a una frecuencia de capitalización dada. Unicamente devuelve el valor numérico. Útil cuando se requieren hacer cálculos adicionales con ese valor.
}
\examples{
monto_compuesto_n(100,12,0.05,12)
}
