% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfuncts.R
\name{tiempo_compuesto_n}
\alias{tiempo_compuesto_n}
\title{Cálculo del tiempo}
\usage{
tiempo_compuesto_n(monto, periodicidad, capital, interes)
}
\arguments{
\item{monto}{Monto esperado a recibir al final del plazo}

\item{periodicidad}{Amplitud de las unidades de tiempo. Ej: 1 equivale a un año, 2 a un semestre, 12 a un mes, 365 a un día.}

\item{capital}{Valor de la inversión inicial}

\item{interes}{Tasa de interés compuesto a la que se dejará la inversión inicial}
}
\value{

}
\description{
Cálculo del tiempo
}
\details{
Función para calcular el tiempo necesario para obtener un monto dado a un cierto tiempo, con una tasa de interés compuesta y capitalización dadas, considerando un capital inicial dado. Unicamente devuelve el valor numérico. Útil cuando se requieren hacer cálculos adicionales con ese valor.
}
\examples{
tiempo_compuesto_n(150,12,200,12)
}
