% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfuncts.R
\name{valor_futuro_simple_n}
\alias{valor_futuro_simple_n}
\title{Cálculo del valor futuro generado a una tasa simple.}
\usage{
valor_futuro_simple_n(capital, tiempo, interes, periodicidad)
}
\arguments{
\item{capital}{Valor de la inversión inicial}

\item{tiempo}{Unidades de tiempo (días, meses, años) en las que se dejará la inversión actual}

\item{interes}{Tasa de interés simple a la que se dejará la inversión inicial}

\item{periodicidad}{Amplitud de las unidades de tiempo. Ej: 1 equivale a un año, 2 a un semestre, 12 a un mes, 365 a un día.}
}
\value{

}
\description{
Cálculo del valor futuro generado a una tasa simple.
}
\details{
Unicamente devuelve el valor numérico. Útil cuando se requieren hacer cálculos adicionales con ese valor.
}
\examples{
valor_futuro_simple_n(100,12,0.05,12)
}
