% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfuncts.R
\name{vfsimple_plot}
\alias{vfsimple_plot}
\title{Gráfico de interés simple}
\usage{
vfsimple_plot(capital, tiempo, interes, periodicidad)
}
\arguments{
\item{capital}{Valor de la inversión inicial}

\item{tiempo}{Unidades de tiempo (días, meses, años) en las que se dejará la inversión actual}

\item{interes}{Tasa de interés simple a la que se dejará la inversión inicial}

\item{periodicidad}{Amplitud de las unidades de tiempo. Ej: 1 equivale a un año, 2 a un semestre, 12 a un mes, 365 a un día.}
}
\value{

}
\description{
Gráfico de interés simple
}
\details{
Generación automática de un gráfico mediante Ggplot2, con el monto obtenido al final del periodo, más el capital inicial. La función además devuelve un dataframe con los montos totales para cada periodo (interés simple).
}
